feat: support files delete

+++ src/articles/articles.service.ts
@@ -6,11 +6,13 @@ import { UpdateArticleInput } from './dto/update-article.input';
 import removeNullishAttrs from 'src/utils/removeNullishAttrs';
 import { Article, ArticleDocument } from './schemas/article.schema';
 import { LangSearchI } from 'src/global/dto/lang-search.input';
+import { FilesService } from 'src/files/files.service';
 
 @Injectable()
 export class ArticlesService {
   constructor(
     @InjectModel(Article.name) private articleModel: Model<ArticleDocument>,
+    private filesService: FilesService,
   ) {}
 
   findAll(titleSearch?: LangSearchI, keywordId?: string) {
@@ -60,7 +62,17 @@ export class ArticlesService {
     return this.articleModel.findOne({ _id }).populate('keywords');
   }
 
+  async deleteArticleFiles(article: Article) {
+    if (!article.thumbnail) return;
+    const thumbnailFilename = article.thumbnail.split('/').slice(-1)[0];
+    await this.filesService.deleteFiles([thumbnailFilename]);
+  }
+
   async delete(_id: string) {
+    const article = await this.articleModel.findOne({ _id });
+    if (!article) return true;
+
+    await this.deleteArticleFiles(article);
     await this.articleModel.deleteOne({ _id });
     return true;
   }
